Fix overflow issue

+++ counting/counting.c
@@ -194,7 +194,7 @@ uint64_t seen_only_once(struct tree_owner *owner) {
 
 uint64_t aggregate_unique_numbers(struct tree_owner ctx[], uint64_t shards) {
 	uint64_t i;
-	int res = 0;
+	uint64_t res = 0;
 	
 	for (i = 0; i < shards; i++)
 		res += ctx[i].elements_in_map;
@@ -204,7 +204,7 @@ uint64_t aggregate_unique_numbers(struct tree_owner ctx[], uint64_t shards) {
 
 uint64_t aggregate_seen_only_once(struct tree_owner ctx[], uint64_t shards) {
 	uint64_t i;
-	int res = 0;
+	uint64_t res = 0;
 	
 	for (i = 0; i < shards; i++) {
 		res += seen_only_once(&ctx[i]);
